+++ src/Core/Dfe.Complete.Domain/Enums/NoteTaskIdentifier.cs
@@ -83,5 +83,117 @@ public enum NoteTaskIdentifier
     
     [Description("master_funding_agreement")]
     [DisplayDescription("Master funding agreement")]
-    MasterFundingAgreement = 20
+    MasterFundingAgreement = 20,
+
+    [Description("confirm_incoming_trust_ceo_contact")]
+    [DisplayDescription("Confirm the incoming trust CEO's details")]
+    ConfirmIncomingTrustCeoContact = 21,
+
+    [Description("risk_protection_arrangement")]
+    [DisplayDescription("Confirm the academy's risk protection arrangements")]
+    ConfirmRiskProtectionArrangements = 22,
+
+    [Description("rpa_policy")]
+    [DisplayDescription("Confirm the academy's risk protection arrangements")]
+    ConfirmRiskProtectionArrangementsPolicy = 23,
+
+    [Description("check_accuracy_of_higher_needs")]
+    [DisplayDescription("Check accuracy of high needs places information")]
+    CheckAccuracyOfHigherNeeds = 24,
+
+    [Description("complete_notification_of_change")]
+    [DisplayDescription("Complete a notification of changes to funded high needs places form")]
+    CompleteNotificationOfChange = 25,
+
+    [Description("conversion_grant")]
+    [DisplayDescription("Process conversion support grant")]
+    ProcessConversionSupportGrant = 26,
+    
+    [Description("sponsored_support_grant")]
+    [DisplayDescription("Confirm and process the sponsored support grant")]
+    ConfirmAndProcessTheSponsoredSupportGrant = 27,
+    
+    [Description("confirm_headteacher_contact")]
+    [DisplayDescription("Confirm the headteacher's details")]
+    ConfirmHeadTeacherDetails = 28,
+
+    [Description("confirm_chair_of_governors_contact")]
+    [DisplayDescription("Confirm the chair of governors' details")]
+    ConfirmChairOfGovernorsDetails = 29,
+
+    [Description("trust_modification_order")]
+    [DisplayDescription("Trust modification order")]
+    TrustModificationOrder = 30,
+
+    [Description("direction_to_transfer")]
+    [DisplayDescription("Direction to transfer")]
+    DirectionToTransfer = 31,
+
+    [Description("one_hundred_and_twenty_five_year_lease")]
+    [DisplayDescription("125 year lease")]
+    OneHundredAndTwentyFiveYearLease = 32,
+
+    [Description("subleases")]
+    [DisplayDescription("Subleases")]
+    Subleases = 33,
+
+    [Description("tenancy_at_will")]
+    [DisplayDescription("Tenancy at will")]
+    TenancyAtWill = 34,
+
+    [Description("school_completed")]
+    [DisplayDescription("Confirm the school has completed all actions")]
+    ConfirmSchoolHasCompletedAllActions = 35,
+
+    [Description("share_information")]
+    [DisplayDescription("Share the information about opening")]
+    ShareInformationAboutOpening = 36,
+
+    [Description("confirm_outgoing_trust_ceo_contact")]
+    [DisplayDescription("Confirm the outgoing trust CEO's details")]
+    ConfirmOutgoingTrustCeoDetails = 37,
+
+    [Description("request_new_urn_and_record")]
+    [DisplayDescription("Request a new URN and record for the academy")]
+    RequestNewUrnAndRecordForAcademy = 38,
+
+    [Description("sponsored_support_grant")]
+    [DisplayDescription("Confirm transfer grant funding level")]
+    ConfirmTransferGrantFundingLevel = 39,
+
+    [Description("check_and_confirm_financial_information")]
+    [DisplayDescription("Check and confirm academy and trust financial information")]
+    CheckAndConfirmAcademyAndTrustFinancialInformation = 40,
+
+    [Description("form_m")]
+    [DisplayDescription("Form M")]
+    FormM = 41,
+
+    [Description("closure_or_transfer_declaration")]
+    [DisplayDescription("Closure or transfer declaration")]
+    ClosureOrTransferDeclaration = 42,
+
+    [Description("bank_details_changing")]
+    [DisplayDescription("Confirm if the bank details for the general annual grant payment need to change")]
+    ConfirmBankDetailsForGeneralAnnualGrantPaymentNeedToChange = 43,
+
+    [Description("confirm_incoming_trust_has_completed_all_actions")]
+    [DisplayDescription("Confirm the incoming trust has completed all actions")]
+    ConfirmIncomingTrustHasCompletedAllActions = 44,
+
+    [Description("land_consent_letter")]
+    [DisplayDescription("Land consent letter")]
+    LandConsentLetter = 45,
+
+    [Description("deed_of_termination_for_the_master_funding_agreement")]
+    [DisplayDescription("Deed of termination for the master funding agreement")]
+    DeedOfTerminationForMasterFundingAgreement = 46,
+
+    [Description("deed_termination_church_agreement")]
+    [DisplayDescription("Deed of termination for the church supplemental agreement")]
+    DeedOfTerminationForChurchSupplementalAgreement = 47,
+
+    [Description("academy_details")]
+    [DisplayDescription("Confirm the academy name")]
+    AcademyDetails = 48
 }