+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/TenancyAtWill/TenancyAtWill.cshtml.cs
@@ -1,0 +1,17 @@
+using Dfe.Complete.Domain.Enums;
+using MediatR;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.TenancyAtWill
+{
+    public class TenancyAtWillModel(ISender sender, IAuthorizationService authorizationService, ILogger<TenancyAtWillModel> logger)
+    : BaseProjectTaskModel(sender, authorizationService, logger, NoteTaskIdentifier.TenancyAtWill)
+    {
+        public override async Task<IActionResult> OnGetAsync()
+        {
+            await base.OnGetAsync();
+            return Page();
+        }
+    }
+}